[@agile-ts/core] added public folders
