+++ packages/core/src/collection/public/index.ts
@@ -1,0 +1,1 @@
+export * from './createCollection';
