+++ packages/core/src/computed/public/index.ts
@@ -1,0 +1,8 @@
+import { CreateAgileSubInstanceInterface } from '../../shared';
+import { CreateComputedConfigInterface } from '../computed';
+
+export * from './createComputed';
+
+export interface CreateComputedConfigInterfaceWithAgile
+  extends CreateAgileSubInstanceInterface,
+    CreateComputedConfigInterface {}
